Make content content-addressable by default
Initially the plan was for content-addressable artifacts opt-in, to
support deduplication and better map to blob storage / OCI storage.

This would allow for the disk-based storage to be plain files in a
directory. However, that did not map well to a digest-based API, where
the disk-based library would suffer performance wise.

Instead, let the disk library be content-addressable and symlink the
files to the blobs instead.